More code clean-up, typing and JIT tags.
More JIT tags have been added where some speed improvements were seen. These were not benchmarked - just a quick execution of the simulation.

Code regarding the movement of the cat has been moved to a separate file. This hopefully will make the physical environment easier to create later on.

Made progress on typing the code. This makes the np arrays easier to track and in some cases was necessary for JIT tags.